Enable syncing Hudi metadata to the Glue Catalog
Hudi has a feature in which it syncs the table's schema and partitions
to the Glue catalog. This is helpful for users who want to query a Hudi
table via AWS Athena.

This commit adds a missing dependency and config settings so the
Hudi/Glue sync now works with this loader, if configured.

Futhermore, this commit forks and amends two Hudi classes:
`AWSGlueCatalogSyncClient` and `GlueCatalogSyncClientConfig`. The
amendments allow us to use a IAM assumed role when connecting to Glue.

The forked classes are licensed under the Apache 2 license.
Our modifications will be contributed back to Hudi via a PR.