Optimizations for larger compute instances
**Increase number of spark writer tasks**. Previously, we did
`.coalesce(1)` on the spark data frame, which meant we could write
really big parquet files. However, on bigger compute instances the data
frame grows so large that writing to storage can become the bottle neck.
After this PR, we do `.coalesce(n)` where `n` scales with the number of
cpu.

**Changes in how windows overlap**. Finalizing a windows comprises two
steps: 1. Union together the pending Data Frames, and 2. write into the
Lake. Previously, we did not start on step 1 until the previous window
had completely finished step 2. After this PR, we start eagerly on step
1 even if the previous window is still on step 2. This behaviour change
is made possible by upgrading common-streams to 0.7.0

**Configurable number of eager windows**. Previously, we let a window
begin processing even when the immediate previous windows was still
finalizing. Now, there is a configuration option so a window can start
even when >1 previous windows are still finalizing. The default value is
to retain the previous behaviour. This is an advanced option that
probably should not be changed; but it is helpful to make it
configurable for some workloads.

**Increased default value of in memory batch size**. Previously, the
loader worked with data frames of 25 MB. After the other changes in this
PR, I found a value of 50 MB works better on big and small compute
instances.