Improve default Delta table properties
This commit changes the default table properties when creating a new table. The changes are only relevant when running the loader for the first time. The new defaults are based on our recent experience of loading to Delta with high even volume.

- `delta.logRetentionDuration`.  This affects users who periodically run a compaction job on their Delta Table. Previously, we kept old log files for 30 days beyond the compaction job.  For high volume pipelines, by reducing this to 1 day we can reduce the number of log files to be managed by Delta.
- `delta.dataSkippingStatsColumns`.  We want Delta to collect stats on columns `load_tstamp`, `collector_tstamp`, `derived_tstamp` and `dvce_created_tstamp`.  Previously we achieving this by moving those four columns to the far left of the table, and then setting `dataSkippingColumns = 4`. Delta has a new option `dataSkippingStatsColumns` where we can explicitly name the columns to index.  It is better for the end user, because they can potentially alter the table to add any custom column to the list.
- `delta.checkpointInterval`. By default, Delta creates a checkpoint every 10 commits.  Because the Lake Loader commits frequently, and because it scales horizontally to multiple loaders, we have found improved efficiency by decreasing how often it writes a checkpoint.

This commit also sets the spark option `spark.databricks.delta.autoCompact.enabled` to false.  This is only needed in case the customer ever manually sets the table property `delta.autoOptimize.autoCompact`: it is important we override the table property.