Trap SIGTERM and start graceful shutdown
This loader tries to do graceful shutdown. Unfortunately, some of the
3rd party libraries try to do their own graceful shutdown as soon as the
JVM begins to shutdown. This is bad for the loader because it interferes
with our delayed graceful shutdown.

This commit works by trapping the SIGTERM so we explicitly start our own
graceful shutdown before the JVM (and therefore 3rd party libs) start
their own shutdown.