Allow disregarding Iglu field's nullability when creating output columns (#66)
Some query engines dislike Iceberg tables in which a STRUCT field is
nullable and a nested field of the STRUCT is non-nullable. For example,
in Snowflake we have seen errors like "SQL execution internal error"
when the nested field contains a null.

This PR adds a config option `respectIgluNullability`.  When set to `false`,
the Lake Loader will make all nested fields nullable. This is slightly
less ideal for data storage, but it brings back compatibility with query
engines like Snowflake.

The default value of the new config option is `true` which maintains the
behaviour of previous versions of the loader.