Bump hudi to 0.15.0 (#70)
As part of this change we also bump spark to 3.5.x when using Hudi. And
bump scala to 2.13.x. Previously we were pinned to earlier versions
because of compatibility with Hudi 0.14.0.

This PR is implemented in a way that we retain the flexibility of
easily supporting a different version of Spark for the Hudi docker
image. I anticipate we might need this flexibility if Iceberg/Delta are
faster to add support for Spark 4.x.