Create table concurrently with subscribing to stream of events (#77)
For KCL apps, we want the KCL to get initialized as early as possible,
so the worker can claim shard leases before they get stolen by other
workers.

Before this PR, the loader initialized the destination table first, and
then subscribed to the stream afterwards.  Initializing the destination
table can be fairly slow, especially because we do things like syncing
to the external catalog, and possibly cleaning up aborted commits.

After this PR, the loader subscribes to the stream concurrently with
initializing the destination table. This lets the KCL claim leases
before they get stolen.