Enable Iceberg's fan-out writer
Previously, we were using the hash Iceberg [write distribution mode][1],
because it is the default for Iceberg. The hash mode does an extra
shuffle before writing. We found that that the write phase of the Lake
Loader was too slow when under high load (Iceberg only): specifically, a
write was taking longer than the loader's "window", and this caused
periods of low cpu usage, where the loader's processing phase was
waiting for the write phase to catch up.

This commit changes to the "none" write distribution mode, and the
fan-out writer.  This removes the need for any spark shuffle task, and
so the loader's write phase should be faster.

This change makes Iceberg writes behave more similar to Delta writes: In
both cases the underlying writer runs a spark task which reads an
un-sorted spark dataframe and fans-out to multiple output partitions in
parallel. It is helpful for Lake Loader development to have a similar
writing model across both formats.

[1]: https://iceberg.apache.org/docs/1.7.1/spark-writes/#writing-distribution-modes