Repartition by event name before writing to the lake
Previously, our Iceberg writer was using the [hash write distribution mode][1]
because that is the default for Iceberg. In this mode, Spark
repartitions by the dataframe immediately before writing to the lake.

After this commit, we explicitly repartition the dataframe as part of
the existing spark task for preparing the final dataframe. This means we
can change the Iceberg write distribution mode to `none`.

Overall this seems to improve the time taken to write a window of events
to Iceberg. This fixes a problem we found, in which the write phase
could get too slow when under high load (Iceberg only): specifically, a
write was taking longer than the loader's "window" and this caused
periods of low cpu usage, where the loader's processing phase was
waiting for the write phase to catch up.

This commit also removes the config option `writerParallelismFraction`.
Before this commit, there were disadvantages to making the writer
parallelism too high, because it would lead to smaller file sizes. But
after this commit, now that we partition by event_name, we might as well
make the writer parallelism as high as reasonably possible, which also
speeds up the write phase of the loader.

Note: this improvement will not help Snowplow users who have changed the
parition key to something different to our default. We might want to
make a follow-up change, in which it auto-discovers the lake's partition
key. For example, some users might want to partition by `app_id` instead
of `event_name`.

[1]: https://iceberg.apache.org/docs/1.7.1/spark-writes/#writing-distribution-modes