Add end-to-end latency metric (#100)
The new metric is called `e2e_latency_millis` and it measures the
difference between an event's `collector_tstamp` and when it committed
to the lake.

The metric is emitted to statsd only on minutes in which some events are
committed to the lake, i.e. the end of a window.  In other minutes
(mid-window) the metric is not defined and not emitted to statsd.

The metric measures the worst-case latency for a window, i.e. latency
for the earliest seen `collector_tstamp`.