Disable AWS SDK deprecation warning (#98)
Newer minor versions of the v1 SDK print a warning to the logs:

```
The AWS SDK for Java 1.x entered maintenance mode starting July 31, 2024 and will reach end of support on December 31, 2025. For more information, see https://aws.amazon.com/blogs/developer/the-aws-sdk-for-java-1-x-is-in-maintenance-mode-effective-july-31-2024/
```

...plus a stack trace.

The Lake Loader uses the v1 SDK via a transitive depedency of Hadoop.
The next release of Hadoop is moving to the v2 SDK, and the Lake Loader
will follow Hadoop latest releases. Therefore it is OK to just disable
the warning message for now.