Revert to unsorted Dataframe for Delta and Hudi (#109)
In #102 I changed how we partition and sort the DataFrame before writing
to the Lake.

The new partitioning is working well for all output formats. But the
extra sort step is not strictly unnecessary for Delta and Hudi. This
commit removes the sort step for Delta/Hudi only, to slightly improve
cpu utilization.