Fix unlikely race condition in LakeWriter
Before #102, we had two mutexes in LakeWriter. In #102 I removed one of
them (amongst other changes).  Now I realize that introduced a potential
race condition, because of how we save, modify, and re-save a dataframe
by name, and because we have parallelism over the that part of the code.

This commit re-introduces the mutex again, to avoid the potential race
condition.

I never observed the race condition, and I think it was very unlikely.
Introducing a mutex here should not impact performance.